requirements.txt
acme==0.31.0
appdirs==1.4.4
asgiref==3.2.10
asn1crypto==0.24.0
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2018.1.18
chardet==3.0.4
ConfigArgParse==0.11.0
configobj==5.0.6
cryptography==2.1.4
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.8
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
filelock==3.0.12
future==0.15.2
idna==2.6
importlib-metadata==1.7.0
importlib-resources==3.0.0
josepy==1.1.0
keyring==10.6.0
keyrings.alt==3.0
mock==2.0.0
ndg-httpsclient==0.4.4
oauthlib==3.1.0
parsedatetime==2.4
pbr==3.1.1
Pillow==7.2.0
psycopg2==2.8.5
pyasn1==0.4.2
pycrypto==2.6.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.3.1
urllib3==1.22
virtualenv==20.0.27
zipp==3.1.0
zope.component==4.3.0
zope.event==4.2.0
